refactor(services): single-source service row-mappers + page→entity shapers
Two residual duplications survived Batch 1's shapers.ts:

(a) Sozluk mapped the same term_summary column selection + TermSummaryRow
literal 3×. Extracted termSummaryColumns + toTermSummaryRow into a sibling
term-summary.ts; getTermSummariesByIds / listTermSummaries /
listTermSummariesConnection reuse them.

(b) The TermPage→Term and PanoPostPage→Post argument maps were restated in
both the read resolver and the delete-refresh. Added toTermFromPage(page) /
toPostFromPage(page, myVote) in shapers.ts so queries.term/mutations.definition.delete
and queries.post/pano-mutations.comment.delete share one mapping.

Co-located the definition/term row shapes with their mappers (definition-row.ts,
term-summary.ts), re-exported from Sozluk.ts, dropping it from 1125 → 989 lines
(under the 1000 gate). Behavior-preserving; wire shapes byte-identical.
